requirements.txt
APScheduler==3.8.1
astroid==2.9.3
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
Deprecated==1.2.13
idna==3.3
iqoptionapi==7.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
platformdirs==2.4.1
pyaes==1.6.1
pyasn1==0.4.8
pylint==2.12.2
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.1.3
requests==2.27.1
rq==1.10.1
rsa==4.8
six==1.16.0
Telethon==1.24.0
toml==0.10.2
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
websocket-client==0.56.0
wrapt==1.13.3
